requirements.txt
tqdm
numpy
opencv-python
onnxruntime-gpu